Refactor code for webpack
