+++ client/src/index.jsx
@@ -1,15 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-
-class App extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
-  render() {
-    return <div>Hello world!!!</div>;
-  }
-}
+import App from './components/App.jsx';
 
 ReactDOM.render(<App />, document.getElementById('app'));
